Fixing
In the original code, we directly referenced the window object in the render method:

maxHeight: window.innerWidth <= 768 ? 'calc(100vh - 10rem)' : 'calc(80vh - 10rem)'

This becomes an issue when the React component is being rendered on the server. Server-Side Rendering (SSR) does not have access to the window object because it's a browser-specific global object. Thus, when the server tries to render this component, it throws an error because it doesn't know what window is.

To solve this, we use a combination of React's useState and useEffect hooks.